refactor(server): inline stack trace collection
