+++ server/src/main/java/org/apache/druid/server/system/module/SystemTableModule.java
@@ -27,8 +27,8 @@ import org.apache.druid.guice.LazySingleton;
 import org.apache.druid.query.SystemTableDataSource;
 import org.apache.druid.server.system.table.ServerPropertiesTableDataProvider;
 import org.apache.druid.server.system.table.ServerPropertiesTableDescriptor;
-import org.apache.druid.server.system.table.StackTraceTableDataProvider;
 import org.apache.druid.server.system.table.StackTraceTableDescriptor;
+import org.apache.druid.server.system.table.SystemStackTraceTableDataProvider;
 import org.apache.druid.server.system.table.SystemTableDataProvider;
 import org.apache.druid.server.system.table.SystemTableDescriptor;
 import org.apache.druid.server.system.table.TaskTableDescriptor;
@@ -62,7 +62,7 @@ public class SystemTableModule implements Module
                       .to(ServerPropertiesTableDataProvider.class)
                       .in(LazySingleton.class);
     dataProviderBinder.addBinding(StackTraceTableDescriptor.TABLE_NAME)
-                      .to(StackTraceTableDataProvider.class)
+                      .to(SystemStackTraceTableDataProvider.class)
                       .in(LazySingleton.class);
   }
 }
